Towards reusable stubs
