+++ touchdown/tests/stubs/aws/volume.py
@@ -1,0 +1,48 @@
+# Copyright 2016 Isotoma Limited
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from .service import ServiceStubber
+
+
+class VolumeStubber(ServiceStubber):
+
+    client_service = 'ec2'
+
+    def add_describe_volumes_empty_response_by_name(self):
+        return self.add_response(
+            'describe_volumes',
+            service_response={
+                'Volumes': [],
+            },
+            expected_params={
+                'Filters': [
+                    {'Name': 'tag:Name', 'Values': [self.resource.name]}
+                ],
+            },
+        )
+
+    def add_describe_volumes_one_response_by_name(self):
+        return self.add_response(
+            'describe_volumes',
+            service_response={
+                'Volumes': [{
+                    'VolumeId': 'vol-abcdef12345',
+                }],
+            },
+            expected_params={
+                'Filters': [
+                    {'Name': 'tag:Name', 'Values': [self.resource.name]}
+                ],
+            },
+        )
